refactor: improve file explorer organization
- Split useFileExplorer into smaller modules:
  - fileOperations.ts: file system operations
  - useContextMenu.ts: context menu logic
  - useDialog.ts: dialog management
  - constants.ts: initial state
- Improved code organization and maintainability
- Added proper TypeScript types
